+++ CHANGELOG.md
@@ -12,6 +12,7 @@ adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) starting v1.
 - Remove dependency: github.com/pkg/errors (#443)
 - Add public Cache.RemainingCost() method
 - Add support for uint keys
+- Implement public Cache.IterValues() method (#475)
 
 **Fixed**